test[skiplog]: extend addon-busy retry to kv-cache and broaden cascade evict
Two follow-up CI runs (25883297200, 25886512750) surfaced three failure
patterns the previous logging-only mitigation did not cover:

- Android `kv-cache-no-system-prompt` was rejected by the registry
  concurrency policy after `kv-cache-stats-verification` hit its 180s
  outer ceiling and left the completion slot occupied. The kv-cache
  executor never went through the busy-retry path.
- Android `logging-all-addons-silent` failed with `Addon stayed busy:
  waited 30000ms` because the prior `logging-persist-across-reload`
  returned `passed: false` (did not throw), so the previous catch-only
  evict never ran and the wedged dep stayed loaded.
- macOS `cancel-then-resume-kv-cache` failed with `expected "4", got
  "5"` on m4 Metal: a 1B model does not handle `2+2` deterministically
  across backends under greedy sampling.

Changes:
- Move `callWhenAddonIdle` / `AddonBusyTimeoutError` /
  `isTransientAddonBusy` into `tests/shared/utils/addon-idle.ts` so
  every executor can absorb the cross-test cascade.
- Wrap kv-cache `runCompletion` in `callWhenAddonIdle` (policy reject
  always fires before the first token, so retry stays safe for both
  streaming and non-streaming paths).
- Broaden `LoggingExecutor.dispatch` to evict the dep on any failure
  (returned `passed: false` or thrown), not only `AddonBusyTimeoutError`.
- Swap the arithmetic prompt in `cancel-then-resume-kv-cache` for an
  alphabet-sequence prompt — a stronger attractor for small models
  across backends.

Verified locally on macOS (local-local-1778827153786) and iOS
(local-local-1778827355368).
