+++ packages/sdk/tests-qvac/tests/shared/executors/logging-executor.ts
@@ -7,11 +7,11 @@ import {
   diffusion,
   unloadModel,
   SDK_LOG_ID,
-  SDK_SERVER_ERROR_CODES,
 } from "@qvac/sdk";
 import { type TestResult } from "@tetherto/qvac-test-suite";
 import { AbstractModelExecutor } from "./abstract-model-executor.js";
 import { loggingTests } from "../../logging-tests.js";
+import { callWhenAddonIdle } from "../utils/addon-idle.js";
 
 // Mirror of `LoggingStreamResponse` from @qvac/sdk (not currently exported).
 export interface LogEntry {
@@ -39,55 +39,6 @@ const INVALID_ID_TIMEOUT_MS = 3_000;
 const DURING_INFERENCE_DRAIN_MS = 1_000;
 const CONCURRENT_DRAIN_MS = 3_000;
 const RELOAD_DRAIN_MS = 5_000;
-const ADDON_BUSY_TIMEOUT_MS = 30_000;
-const ADDON_BUSY_POLL_MS = 250;
-
-// Documented busy throw from infer-llamacpp-llm; we retry until idle.
-const ADDON_BUSY_MARKER = "a job is already set or being processed";
-
-class AddonBusyTimeoutError extends Error {
-  constructor(timeoutMs: number, cause: unknown) {
-    super(`Addon stayed busy: waited ${timeoutMs}ms`, { cause });
-    this.name = "AddonBusyTimeoutError";
-  }
-}
-
-// "Slot still occupied" surfaces as either the legacy addon throw or, since
-// the `oneAtATimePerModel` registry policy, a typed admission rejection.
-// Errors cross the RPC boundary as `RPCError`, so we match by code, not
-// `instanceof`.
-function isTransientAddonBusy(err: unknown): boolean {
-  if (
-    err !== null
-    && typeof err === "object"
-    && "code" in err
-    && (err as { code?: unknown }).code === SDK_SERVER_ERROR_CODES.REQUEST_REJECTED_BY_POLICY
-  ) {
-    return true;
-  }
-  const msg = err instanceof Error ? err.message : String(err);
-  return msg.includes(ADDON_BUSY_MARKER);
-}
-
-export async function callWhenAddonIdle<T>(
-  fn: () => Promise<T>,
-  timeoutMs = ADDON_BUSY_TIMEOUT_MS,
-  intervalMs = ADDON_BUSY_POLL_MS,
-): Promise<T> {
-  const deadline = Date.now() + timeoutMs;
-  while (true) {
-    try {
-      return await fn();
-    } catch (err) {
-      if (isTransientAddonBusy(err)) {
-        if (Date.now() >= deadline) throw new AddonBusyTimeoutError(timeoutMs, err);
-        await sleep(intervalMs);
-        continue;
-      }
-      throw err;
-    }
-  }
-}
 
 const TRIGGER_KEYS = [
   "llm", "embed", "tts", "nmt", "diffusion",
@@ -106,6 +57,8 @@ const HANDLER_KEYS = [
   "concurrent",
   "reload",
 ] as const;
+
+type HandlerKey = typeof HANDLER_KEYS[number];
 
 const isHandlerKey = makeKeyGuard(HANDLER_KEYS);
 
@@ -151,20 +104,27 @@ export class LoggingExecutor extends AbstractModelExecutor<typeof loggingTests> 
       throw new Error(`Test "${testId}" has missing/invalid params.handler: ${String(handler)}`);
     }
 
+    let result: TestResult;
     try {
-      switch (handler) {
-        case "addon-logging":     return await this.runAddonLogging(testId, params);
-        case "invalid-model-id":  return await this.runInvalidModelId(requireInvalidModelIdParams(params));
-        case "during-inference":  return await this.runDuringInference(testId, params as DuringInferenceParams);
-        case "concurrent":        return await this.runConcurrent(testId, requireOperationsParams(params));
-        case "reload":            return await this.runReload(testId);
-      }
+      result = await this.runHandler(handler, testId, params);
     } catch (error) {
-      // Break the busy cascade: evict the wedged dep so the next test reloads a fresh model.
-      if (error instanceof AddonBusyTimeoutError) {
-        await this.evictTestDeps(testId);
-      }
-      return wrapError(testId, error);
+      result = wrapError(testId, error);
+    }
+
+    // Break the busy cascade: evict the dep on any failure so the next test reloads a fresh model.
+    if (!result.passed) {
+      await this.evictTestDeps(testId);
+    }
+    return result;
+  }
+
+  private runHandler(handler: HandlerKey, testId: string, params: AnyParams): Promise<TestResult> {
+    switch (handler) {
+      case "addon-logging":     return this.runAddonLogging(testId, params);
+      case "invalid-model-id":  return this.runInvalidModelId(requireInvalidModelIdParams(params));
+      case "during-inference":  return this.runDuringInference(testId, params as DuringInferenceParams);
+      case "concurrent":        return this.runConcurrent(testId, requireOperationsParams(params));
+      case "reload":            return this.runReload(testId);
     }
   }
 
@@ -174,7 +134,7 @@ export class LoggingExecutor extends AbstractModelExecutor<typeof loggingTests> 
     try {
       await this.resources.evict(dep);
     } catch (error) {
-      console.warn(`LoggingExecutor: failed to evict dep "${dep}" after AddonBusyTimeoutError: ${error}`);
+      console.warn(`LoggingExecutor: failed to evict dep "${dep}" after failure: ${error}`);
     }
   }
 
@@ -280,9 +240,7 @@ export class LoggingExecutor extends AbstractModelExecutor<typeof loggingTests> 
 
     const reloadedModelId = await this.resources.ensureLoaded(dep);
 
-    // Join the trigger before returning — `collectLogs` races it against
-    // log arrival, so without this the next test can race the still-open
-    // completion slot.
+    // Join the trigger before returning so the next test cannot race the open completion slot.
     let triggerPromise: Promise<void> = Promise.resolve();
     const result = await collectLogs({
       testId,
